chore: made type human-readable for GitHub comment
